requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
bcrypt==4.1.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
passlib==1.7.4
pyotp==2.9.0
qrcode[pil]==7.4.2
cryptography>=41.0.0
emails==0.6
itsdangerous==2.1.2
aiosmtplib==3.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2